+++ example/lib/main.dart
@@ -3,14 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:ultimate_extension/ultimate_extension.dart';
 
 void main() async {
-  final result = await compute(1000).logExecution("Compute");
-  print("Result: $result");
   runApp(MaterialApp(home: ExamplePage()));
-}
-
-Future<int> compute(int data) async {
-  await Future.delayed(Duration(seconds: 1));
-  return data;
 }
 
 class ExamplePage extends StatelessWidget {
@@ -28,7 +21,8 @@ class ExamplePage extends StatelessWidget {
             ),
             Builder(builder: (context) {
               return Container(
-                decoration: BoxDecoration(color: Colors.red),
+                decoration:
+                    BoxDecoration(color: Colors.red, shape: BoxShape.circle),
                 width: 100,
                 height: 100,
               ).shimmer(isLoading: true);